refactor: 관련 훅들 hooks로 이동
