+++ src/components/Snackbar/index.ts
@@ -1,3 +1,3 @@
 export { SnackbarProvider } from './SnackbarProvider';
-export { useSnackbar } from './useSnackbar';
+export { useSnackbar } from './hooks/useSnackbar';
 export type * from './Snackbar.type';
