+++ src/components/index.ts
@@ -56,5 +56,5 @@ export { Divider } from './Divider';
 export type { DividerProps } from './Divider';
 
 export { SnackbarProvider } from './Snackbar/SnackbarProvider';
-export { useSnackbar } from './Snackbar/useSnackbar';
+export { useSnackbar } from './Snackbar/hooks/useSnackbar';
 export type * from './Snackbar';
